规格参数增删改查

+++ mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/SpecificationServiceImpl.java
@@ -4,11 +4,15 @@ import com.alibaba.fastjson.JSONObject;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.SpecGroupDTO;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.SpecificationService;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 import tk.mybatis.mapper.entity.Example;
 
@@ -29,7 +33,11 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Resource
     private SpecGroupMapper specGroupMapper;
 
-    //查询
+
+    @Resource
+    private SpecParamMapper specParamMapper;
+
+    //查询//品牌
     @Override
     public Result<List<SpecGroupEntity>> getSepcGroupInfo(SpecGroupDTO specGroupDTO) {
 
@@ -43,6 +51,7 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
 
     //新增
     @Override
+    @Transactional
     public Result<JSONObject> saveSpecGroup(SpecGroupDTO specGroupDTO) {
 
         specGroupMapper.insertSelective(BaiduBeanUtil.copyProperties(specGroupDTO,SpecGroupEntity.class));
@@ -60,9 +69,61 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
 
     //删除
     @Override
+    @Transactional
     public Result<JSONObject> deleteSpecGroupById(Integer id) {
 
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity>specParamEntities=specParamMapper.selectByExample(example);
+        if (specParamEntities.size()>0){
+            return this.setResultSuccess("该规格组有数据无法删除");
+        }
+
         specGroupMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
+
+
+    //------------------------------------------------------------------------------------------规格参数
+    //通过条件查询规格参数
+    @Override
+    @Transactional
+    public Result<List<SpecParamEntity>> getSpecParamInfo(SpecParamDTO specParamDTO) {
+
+        SpecParamEntity specParamEntity = BaiduBeanUtil.copyProperties(specParamDTO, SpecParamEntity.class);
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",specParamEntity.getGroupId());
+
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+
+        return this.setResultSuccess(specParamEntities);
+    }
+
+    //新增
+    @Override
+    @Transactional
+    public Result<JSONObject> saveSpecParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.insertSelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+
+        return this.setResultSuccess();
+    }
+
+    //修改
+    @Override
+    @Transactional
+    public Result<JSONObject> editSpecParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.updateByPrimaryKeySelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    //删除
+    @Override
+    @Transactional
+    public Result<JSONObject> deleteSpecParam(Integer id) {
+
+        specParamMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
